# ios/CanonicalAPI/Sources/CanonicalAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Money Monitor canonical API",
    "version": "1",
    "description": "The Mac-authoritative, schema-generated contract shared by the Mac-local and paired-iPhone listeners."
  },
  "servers": [{ "url": "http://127.0.0.1" }],
  "paths": {
    "/api/v1/categories": {
      "get": {
        "operationId": "listCategories",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "List canonical categories",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoryListResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCategory",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Create a canonical category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Accepted create or receipt replay",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CategoryResponse" } }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{id}": {
      "patch": {
        "operationId": "updateCategory",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Update a canonical category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CategoryResponse" } }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCategory",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Delete a canonical category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "expectedVersion",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoryDeleteResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/home": {
      "get": {
        "operationId": "getHomeOverview",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Read the Mac-calculated Home overview projection",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HomeOverviewResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference": {
      "get": {
        "operationId": "getReference",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Read the canonical foundation resource",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReferenceResponse" } }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/{id}": {
      "patch": {
        "operationId": "updateReference",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Update the canonical foundation resource",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReferenceUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReferenceResponse" } }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteReference",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Delete the canonical foundation resource",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          {
            "name": "expectedVersion",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReferenceDeleteResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/commands/refresh": {
      "post": {
        "operationId": "requestReferenceRefresh",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "shared",
        "summary": "Request a receipt-protected refresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReferenceCommandRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReferenceCommandResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics": {
      "get": {
        "operationId": "getDiagnostics",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "macOnly",
        "summary": "Read trusted Mac diagnostics",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DiagnosticsResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/pairing/status": {
      "get": {
        "operationId": "getPairingStatus",
        "security": [{ "bearerAuth": [] }],
        "x-route-policy": "pairing",
        "summary": "Read paired-device status",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PairingStatusResponse" }
              }
            }
          },
          "4XX": {
            "description": "Stable coded error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CanonicalErrorEnvelope" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
...
      "CategoryListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "maxItems": 500,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer", "exclusiveMinimum": 0, "maximum": 9007199254740991 },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                },
                "label": { "type": "string", "minLength": 1, "maxLength": 80 },
                "color": { "type": ["string", "null"], "pattern": "^#[0-9a-fA-F]{6}$" },
                "rules": { "type": ["string", "null"], "maxLength": 2000 },
                "defaultOwnerType": {
                  "type": "string",
                  "enum": ["member", "shared", "unassigned"]
                },
                "defaultOwnerMemberId": {
                  "type": ["integer", "null"],
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "ignoredFromStats": { "type": "boolean" },
                "resourceVersion": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                }
              },
              "required": [
                "id",
                "name",
                "label",
                "color",
                "rules",
                "defaultOwnerType",
                "defaultOwnerMemberId",
                "ignoredFromStats",
                "resourceVersion",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "apiVersion": { "type": "string", "const": "1" },
              "generatedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "source": { "type": "string", "const": "mac-authoritative" },
              "calculationVersion": { "type": "string", "minLength": 1, "maxLength": 64 },
              "completeness": { "type": "string", "enum": ["complete", "partial"] },
              "estimated": { "type": "boolean" },
              "resourceVersion": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
              "refreshHints": {
                "maxItems": 20,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "domain": { "type": "string", "minLength": 1, "maxLength": 64 },
                    "resourceIds": {
                      "maxItems": 100,
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      }
                    }
                  },
                  "required": ["domain", "resourceIds"],
                  "additionalProperties": false
                }
              },
              "missingSections": {
                "maxItems": 20,
                "type": "array",
                "items": { "type": "string", "minLength": 1, "maxLength": 64 }
              },
              "receipt": {
                "type": "object",
                "properties": {
                  "idempotencyKey": { "type": "string", "minLength": 1, "maxLength": 128 },
                  "replayed": { "type": "boolean" }
                },
                "required": ["idempotencyKey", "replayed"],
                "additionalProperties": false
              },
              "ownerMembers": {
                "maxItems": 100,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "exclusiveMinimum": 0, "maximum": 9007199254740991 },
                    "name": { "type": "string", "minLength": 1, "maxLength": 120 }
                  },
                  "required": ["id", "name"],
                  "additionalProperties": false
                }
              }
            },
            "required": ["apiVersion", "generatedAt", "source", "ownerMembers"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false
      },
...
      "ReferenceUpdateRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "expectedVersion": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
          "title": { "type": "string", "minLength": 1, "maxLength": 120 },
          "amount": {
            "type": "object",
            "properties": {
              "value": { "type": "string", "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d{1,4})?$" },
              "currencyCode": { "type": "string", "pattern": "^[A-Z]{3}$" }
            },
            "required": ["value", "currencyCode"],
            "additionalProperties": false
          }
        },
        "required": ["expectedVersion"],
        "additionalProperties": false
      },
...
    }
  }
}